Implement electronic property model.

+++ src/pypolymlp/core/interface_yaml.py
@@ -42,8 +42,8 @@ def parse_sscha_yamls(yamlfiles: list[str]):
 def parse_structure_from_yaml(yml_data: yaml, tag: str = "unitcell"):
     """Parse structure in yaml data."""
     cell_dict = yml_data[tag]
-    cell_dict["axis"] = np.array(cell_dict["axis"]).T
-    cell_dict["positions"] = np.array(cell_dict["positions"]).T
+    cell_dict["axis"] = np.array(cell_dict["axis"], dtype=float).T
+    cell_dict["positions"] = np.array(cell_dict["positions"], dtype=float).T
     cell = PolymlpStructure(**cell_dict)
     return cell
